requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==23.2.0
    # via
    #   outcome
    #   trio
certifi==2024.2.2
    # via selenium
cffi==1.16.0
    # via trio
h11==0.14.0
    # via wsproto
idna==3.6
    # via trio
outcome==1.3.0.post0
    # via trio
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pysocks==1.7.1
    # via urllib3
selenium==4.18.1
    # via -r requirements.in
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.24.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.10.0
    # via selenium
urllib3[socks]==2.2.1
    # via selenium
wsproto==1.2.0
    # via trio-websocket
